requirements.txt
amqp==5.0.6
attrs==20.3.0
bidict==0.21.2
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-datauri==1.0.0
python-dotenv==0.17.0
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
tensorflow-datasets==4.2.0
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
